Move CodeRenderer
