+++ src/components/HTMLEngineProvider/BaseHTMLEngineProvider.js
@@ -5,15 +5,13 @@ import {
     TRenderEngineProvider,
     RenderHTMLConfigProvider,
     defaultHTMLElementModels,
-    splitBoxModelStyle,
 } from 'react-native-render-html';
 import PropTypes from 'prop-types';
 import AnchorRenderer from './HTMLRenderers/AnchorRenderer';
+import CodeRenderer from './HTMLRenderers/CodeRenderer';
 import Config from '../../CONFIG';
 import styles from '../../styles/styles';
-import * as StyleUtils from '../../styles/StyleUtils';
 import fontFamily from '../../styles/fontFamily';
-import InlineCodeBlock from '../InlineCodeBlock';
 import AttachmentModal from '../AttachmentModal';
 import ThumbnailImage from '../ThumbnailImage';
 import variables from '../../styles/variables';
@@ -60,40 +58,6 @@ function computeEmbeddedMaxWidth(tagName, contentWidth) {
         return Math.min(MAX_IMG_DIMENSIONS, contentWidth);
     }
     return contentWidth;
-}
-
-function CodeRenderer(props) {
-    // We split wrapper and inner styles
-    // "boxModelStyle" corresponds to border, margin, padding and backgroundColor
-    const {boxModelStyle, otherStyle: textStyle} = splitBoxModelStyle(props.style);
-
-    // Get the correct fontFamily variant based in the fontStyle and fontWeight
-    const font = StyleUtils.getFontFamilyMonospace({
-        fontStyle: textStyle.fontStyle,
-        fontWeight: textStyle.fontWeight,
-    });
-
-    const textStyleOverride = {
-        fontFamily: font,
-
-        // We need to override this properties bellow that was defined in `textStyle`
-        // Because by default the `react-native-render-html` add a style in the elements,
-        // for example the <strong> tag has a fontWeight: "bold" and in the android it break the font
-        fontWeight: undefined,
-        fontStyle: undefined,
-    };
-
-    const defaultRendererProps = _.omit(props, ['TDefaultRenderer', 'style']);
-
-    return (
-        <InlineCodeBlock
-            defaultRendererProps={defaultRendererProps}
-            TDefaultRenderer={props.TDefaultRenderer}
-            boxModelStyle={boxModelStyle}
-            textStyle={{...textStyle, ...textStyleOverride}}
-            key={props.key}
-        />
-    );
 }
 
 function EditedRenderer(props) {
