+++ src/components/HTMLEngineProvider/HTMLRenderers/htmlRendererPropTypes.js
@@ -2,4 +2,7 @@ import PropTypes from 'prop-types';
 
 export default {
     tnode: PropTypes.object,
+    TDefaultRenderer: PropTypes.object,
+    key: PropTypes.string,
+    style: PropTypes.object,
 };
